+++ packages-internal/scripts/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.1
+
+- Unpinned version of the @mui-internal/docs-utils dependency.
+- Corrected the README file.
+
 ## 1.0.0
 
 Initial release as an npm package.